fix: resolve 13 audit issues — append bug, errcheck, errorlint, gofmt
- Fix data-loss bug in rpc_blockchain.go where append results were discarded
- Check resp.Body.Close and db.Close errors
- Use %w for proper error wrapping in network/rpc*.go
- Fix gofmt in tx/batch.go and x402/x402_test.go
- Remove unnecessary uint64 conversions in x402/htlc_tx.go

+++ x402/htlc_tx.go
@@ -278,7 +278,7 @@ func BuildSellerClaimTx(params *SellerClaimParams) (*transaction.Transaction, er
 	}
 
 	// Estimate claim tx size: ~10 overhead + ~(73+33+32+1) unlocking + script + ~40 output.
-	estSize := uint64(10 + 73 + 33 + 32 + 1 + uint64(len(params.HTLCScript)) + 40)
+	estSize := 10 + 73 + 33 + 32 + 1 + uint64(len(params.HTLCScript)) + 40
 	estFee := estSize * feeRate
 
 	if params.FundingAmount <= estFee {
@@ -383,7 +383,7 @@ func BuildSellerPreSignedRefund(params *SellerPreSignParams) (*SellerPreSignResu
 
 	// Estimate refund tx size: ~10 overhead + ~(1 + 73 + 73 + 1) unlocking
 	// (OP_0 + two sigs + OP_FALSE) + script + ~40 output.
-	estSize := uint64(10 + 1 + 73 + 73 + 1 + uint64(len(params.HTLCScript)) + 40)
+	estSize := 10 + 1 + 73 + 73 + 1 + uint64(len(params.HTLCScript)) + 40
 	estFee := estSize * feeRate
 
 	if params.FundingAmount <= estFee {
